controle_acoes/requirements_desktop.txt
# Dependências adicionais para a versão desktop
# Instale com:  pip install -r requirements_desktop.txt

pywebview>=4.0
